add some features

+++ backend/app/restaurant.py
@@ -1,10 +1,9 @@
 from flask import Blueprint, request, jsonify
 from .config import Session
-from .models import Restaurant
+from .models import Restaurant, User
 from .auth import token_required
 
 restaurant_bp = Blueprint('restaurant', __name__)
-
 
 @restaurant_bp.route('/restaurant', methods=['POST'])
 @token_required
@@ -12,16 +11,13 @@ def add_restaurant(current_user):
     data = request.get_json()
     session = Session()
     new_restaurant = Restaurant(
-        nama_restoran=data['nama_restoran'],
-        alamat=data['alamat'],
-        nomor_telepon=data['nomor_telepon'],
+        user_id=current_user.user_id,
         jam_operasional=data['jam_operasional']
     )
     session.add(new_restaurant)
     session.commit()
     session.close()
     return jsonify({'message': 'New restaurant added!'})
-
 
 @restaurant_bp.route('/restaurant/<int:restaurant_id>', methods=['PUT'])
 @token_required
@@ -32,14 +28,13 @@ def update_restaurant(current_user, restaurant_id):
     if not restaurant:
         session.close()
         return jsonify({'message': 'Restaurant not found!'}), 404
-    restaurant.nama_restoran = data['nama_restoran']
-    restaurant.alamat = data['alamat']
-    restaurant.nomor_telepon = data['nomor_telepon']
+    if restaurant.user_id != current_user.user_id:
+        session.close()
+        return jsonify({'message': 'Unauthorized to update this restaurant!'}), 403
     restaurant.jam_operasional = data['jam_operasional']
     session.commit()
     session.close()
     return jsonify({'message': 'Restaurant updated!'})
-
 
 @restaurant_bp.route('/restaurant/<int:restaurant_id>', methods=['DELETE'])
 @token_required
@@ -49,11 +44,13 @@ def delete_restaurant(current_user, restaurant_id):
     if not restaurant:
         session.close()
         return jsonify({'message': 'Restaurant not found!'}), 404
+    if restaurant.user_id != current_user.user_id:
+        session.close()
+        return jsonify({'message': 'Unauthorized to delete this restaurant!'}), 403
     session.delete(restaurant)
     session.commit()
     session.close()
     return jsonify({'message': 'Restaurant deleted!'})
-
 
 @restaurant_bp.route('/restaurant', methods=['GET'])
 def get_restaurants():
@@ -61,10 +58,24 @@ def get_restaurants():
     restaurants = session.query(Restaurant).all()
     result = [{
         'restaurant_id': res.restaurant_id,
-        'nama_restoran': res.nama_restoran,
-        'alamat': res.alamat,
-        'nomor_telepon': res.nomor_telepon,
+        'nama_restoran': res.user.nama,
+        'alamat': res.user.alamat,
+        'nomor_telepon': res.user.nomor_telepon,
         'jam_operasional': res.jam_operasional
     } for res in restaurants]
+    session.close()
+    return jsonify(result)
+
+@restaurant_bp.route('/restaurant/<int:user_id>', methods=['GET'])
+def get_restaurant(user_id):
+    session = Session()
+    res= session.query(Restaurant).filter_by(user_id=user_id).first()
+    result = {
+        'restaurant_id': res.restaurant_id,
+        'nama_restoran': res.user.nama,
+        'alamat': res.user.alamat,
+        'nomor_telepon': res.user.nomor_telepon,
+        'jam_operasional': res.jam_operasional
+    }
     session.close()
     return jsonify(result)
